ChartVision-v1.0 3/requirements.txt
anthropic>=0.39.0
mss>=9.0.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
